+++ toontown/compiler/clashdna/dna/components/DNANode.py
@@ -24,9 +24,9 @@ class DNANode(DNAGroup):
         packer = DNAGroup.traverse(self, recursive=False, verbose=verbose)
         packer.name = 'DNANode'  # Override the name for debugging.
         for component in self.pos:
-            packer.pack('position', int(component * 100), INT32)
+            packer.pack('position', int(component * 100), FLOAT64)
         for component in self.hpr:
-            packer.pack('rotation', int(component * 100), INT32)
+            packer.pack('rotation', int(component * 100), FLOAT64)
         for component in self.scale:
             packer.pack('scale', int(component * 100), UINT16)
         if recursive:
